Chapter09/myblog/requirements.txt
Django
djangorestframework
psycopg2-binary
python-social-auth
redis
django-cacheops
factoryboy